chore: code files structure adjust
